extra controller methods for committing an order along with authorization for editing products

+++ backend/app/Http/Controllers/ProductController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
 use App\Models\Product;
+use App\Models\ProductOrder;
 use http\Env\Response;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
@@ -23,8 +24,24 @@ class ProductController extends Controller
      */
     public function store(StoreProductRequest $request)
     {
+        if (is_null($request->user()) || $request->user()->cannot('create', Product::class)) {
+            abort(403);
+        }
         $product = Product::create($request->validated());
         return response()->json($product);
+    }
+
+    public function makeOrder(Product $product, ProductOrder $productOrder) {
+        if ($productOrder->quantity > $product->stock) {
+            return response()->json(["error" => "Insufficient stock"], 400);
+        }
+
+        $product->stock -= $productOrder->quantity;
+        $productOrder->is_committed = true;
+        $product->save();
+        $productOrder->save();
+
+        return response()->json($productOrder);
     }
 
     /**
@@ -36,7 +53,7 @@ class ProductController extends Controller
            $product = Product::findOrFail($id);
            return response()->json($product);
         } catch (\Throwable $th) {
-            return response()->status(204);
+            abort(404);
         }
     }
 
@@ -45,24 +62,30 @@ class ProductController extends Controller
      */
     public function update(UpdateProductRequest $request, $id)
     {
+        if (is_null($request->user()) || $request->user()->cannot('update', Product::class)) {
+            abort(403);
+        }
         try {
             $product = Product::findOrFail($id);
             $product->update($request->validated());
             return response()->json($product);
         } catch (ModelNotFoundException $e) {
-            return response()->status(204);
+            abort(404);
         }
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($id)
+    public function destroy($request, $id)
     {
+        if (is_null($request->user()) || $request->user()->cannot('delete', Product::class)) {
+            abort(403);
+        }
         try {
             Product::destroy($id);
         } catch (\Throwable $th) {
-            return response()->status(204);
+            abort(404);
         }
     }
 }
